requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
django-rest-framework-nested==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
gunicorn==23.0.0
importlib_resources==6.4.4
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
phonenumbers==8.13.42
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==74.1.1
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
